Optimize iterator performance: TForCall inline, upvalue_ptrs cache, CallInfo compression
Three optimizations targeting Lua-closure-based iterator performance:

1. TForCall inline Lua closure call path: Added inline fast path for Lua
   function iterators in TForCall handler, avoiding handle_call/FrameAction
   overhead. Gives +12% for custom stateless iterators.

2. Cache upvalue_ptrs in CallInfo: Store resolved upvalue array pointer
   directly in CallInfo, avoiding the func->GcPtr->GcRClosure->LuaFunction->
   UpvalueStore enum match chain on every GetUpval/SetUpval (saves 2-3 loads
   + 1 branch per access). Gives +15% for closure-based iterators.

3. CallInfo field compression: Changed func_offset and top from usize to u32
   to keep CallInfo at 72 bytes after adding upvalue_ptrs (8 bytes). Without
   compression, the struct grew to 80 bytes causing cache regression on
   ipairs/pairs. With compression, all C-function-based iterators maintain
   baseline performance.

Results (vs baseline):
- Custom stateless iter: +12% (30.7K -> 34.2K)
- Closure iterator:     +15% (201K -> 231K)
- Multi-value iterator: +15% (14.7K -> 16.9K)
- ipairs/pairs/next:    no regression (within noise)
- Arithmetic/functions: no regression

Added compile-time assertion: CallInfo must remain 72 bytes.

+++ crates/luars/src/lua_vm/execute/closure_handler.rs
@@ -50,7 +50,7 @@ pub fn handle_closure(
     lua_state.set_top_raw(new_top);
     lua_state.check_gc()?;
 
-    let frame_top = lua_state.get_call_info(frame_idx).top;
+    let frame_top = lua_state.get_call_info(frame_idx).top as usize;
     lua_state.set_top_raw(frame_top);
     Ok(())
 }
